6408445 logadm rotates regardless with -s 2g

+++ usr/src/cmd/logadm/fn.h
@@ -2,9 +2,8 @@
  * CDDL HEADER START
  *
  * The contents of this file are subject to the terms of the
- * Common Development and Distribution License, Version 1.0 only
- * (the "License").  You may not use this file except in compliance
- * with the License.
+ * Common Development and Distribution License (the "License").
+ * You may not use this file except in compliance with the License.
  *
  * You can obtain a copy of the license at usr/src/OPENSOLARIS.LICENSE
  * or http://www.opensolaris.org/os/licensing.
@@ -20,8 +19,8 @@
  * CDDL HEADER END
  */
 /*
- * Copyright (c) 2001 by Sun Microsystems, Inc.
- * All rights reserved.
+ * Copyright 2006 Sun Microsystems, Inc.  All rights reserved.
+ * Use is subject to license terms.
  *
  * logadm/fn.h -- public definitions for fn module
  */
@@ -65,7 +64,7 @@ void fn_list_addfn_list(struct fn_list *fnlp, struct fn_list *fnlp2);
 void fn_list_appendrange(struct fn_list *fnlp,
     const char *s, const char *limit);
 void fn_list_print(struct fn_list *fnlp, FILE *stream);
-size_t fn_list_totalsize(struct fn_list *fnlp);
+off_t fn_list_totalsize(struct fn_list *fnlp);
 struct fn *fn_list_popoldest(struct fn_list *fnlp);
 boolean_t fn_list_empty(struct fn_list *fnlp);
 int fn_list_count(struct fn_list *fnlp);
